detectability report
