+++ src/dlomix/reports/__init__.py
@@ -1,7 +1,9 @@
+from .DetectabilityReport import DetectabilityReport
 from .IntensityReport import IntensityReport
 from .RetentionTimeReport import RetentionTimeReport
 
 __all__ = [
     "RetentionTimeReport",
     "IntensityReport",
+    "DetectabilityReport",
 ]
